Added Admin users
Made it so that the first user to sign up becomes an Admin, so that they can add more users, but also made sure that non-Admins can't add or delete users, only view the user list and their profile.

While adding new users, you can decide whether or not they should be an Admin.

Also had to change the user table in the database to include the is_admin column.

+++ main.py
@@ -8,7 +8,7 @@ Created on Thu Aug 28 23:15:07 2025
 
 import sqlite3
 
-login = []
+login = [] # user data, in this order: username, email, user_id, is_admin
 
 def sign_in():
     conn = sqlite3.connect("user_db")
@@ -22,7 +22,7 @@ def sign_in():
         print("\nThere are no signed up users.\n")
     
     else:
-        email = input("E-mail: ")
+        email = input("\nE-mail: ")
         password = input("Password: ")
         
         # checking if the user exists in the database and if the credentials are correct
@@ -30,9 +30,11 @@ def sign_in():
             if row[2] == password and row[3] == email:
                 username = row[1]
                 user_id = row[0]
+                is_admin = row[4]
                 login.append(username)
                 login.append(email)
                 login.append(user_id)
+                login.append(is_admin)
                 break
         if not login:
             print("\nUser not found. Please try again.\n")
@@ -40,31 +42,49 @@ def sign_in():
     conn.close()
 
 def add_user():
-    username = input("Username: ")
-    email = input("E-mail: ")
-    password = input("Password: ")
-    
-    conn = sqlite3.connect("user_db")
-    cur = conn.cursor()
-    
-    cur.execute("SELECT user_id FROM user ORDER BY user_id")
-    
-    result = cur.fetchall()
-    
-    new_id = 1
-    # creating a new id for the user, so that they don't all have the same id
-    for i in result:
-        if new_id == i[0]:
-            new_id += 1
-        else:
-            break
-    print(new_id)
-    cur.execute(f"INSERT INTO user VALUES ({new_id}, '{username}', '{password}', '{email}')")
-    conn.commit()
-    
-    conn.close()
-    
-    print(f"\nUser {username} has been added.\n")
+    if login[3] == "yes": # if the user is an Admin
+        username = input("Username: ")
+        email = input("E-mail: ")
+        password = input("Password: ")
+        
+        conn = sqlite3.connect("user_db")
+        cur = conn.cursor()
+        
+        cur.execute("SELECT user_id FROM user ORDER BY user_id")
+        
+        result = cur.fetchall()
+        
+        new_id = 1
+        # creating a new id for the user, so that they don't all have the same id
+        for i in result:
+            if new_id == i[0]:
+                new_id += 1
+            else:
+                break
+        print(new_id)
+        if not result:
+            cur.execute(f"INSERT INTO user VALUES ({new_id}, '{username}', '{password}', '{email}', 'yes')")
+        choice = ""
+        
+        while choice != "cancel":
+            choice = input("\nShould this user be an Admin? (yes/no): ")
+            match(choice):
+                case "yes":
+                    cur.execute(f"INSERT INTO user VALUES ({new_id}, '{username}', '{password}', '{email}', 'yes')")
+                    conn.commit()
+                    break
+                case "no":
+                    cur.execute(f"INSERT INTO user VALUES ({new_id}, '{username}', '{password}', '{email}', 'no')")
+                    conn.commit()
+                    break
+                case _:
+                    print("\nInvalid option.\n")
+        
+        conn.close()
+        
+        print(f"\nUser {username} has been added.\n")
+    else:
+        print("\nOnly Admins can add or delete users.\n")
 
 def view_all_users():
     print("\nUsers:")
@@ -73,7 +93,7 @@ def view_all_users():
     cur = conn.cursor()
     
     for row in cur.execute("SELECT * FROM user ORDER BY user_id"):
-        print(f"ID: {row[0]}, username: {row[1]}")
+        print(f"ID: {row[0]}, username: {row[1]}, Admin: {row[4]}")
     
     conn.close()
     
@@ -81,72 +101,75 @@ def view_all_users():
 choice = 0
 
 def delete_user():
-    conn = sqlite3.connect("user_db")
-    
-    result = conn.execute("SELECT * FROM user ORDER BY user_id")
-    
-    print("\nUsers:")
-    
-    # display the users in the database
-    for row in result:
-        print(f"ID: {row[0]}, username: {row[1]}")
-    
-    conn.close()
-    
-    user_id = 0
-    
-    while user_id != -1: # -1 is used to cancel the operation.
-        user_found = False
-        
+    if login[3] == "yes": # if the user is an Admin
         conn = sqlite3.connect("user_db")
-        cur = conn.cursor()
         
-        result = conn.execute("SELECT * FROM user")
+        result = conn.execute("SELECT * FROM user ORDER BY user_id")
         
-        user_id = int(input("\nInsert the ID of the user you'd like to remove. Insert -1 to cancel."))
+        print("\nUsers:")
         
-        if user_id == -1:
-            break
-        
-        # finding the user
+        # display the users in the database
         for row in result:
-            x = 0
-            
-            if user_id == row[x]:
-                user_found = True
-                break
-            x += 1
-            
-        if user_found:
-            result = cur.execute(f"SELECT username FROM user WHERE user_id = {user_id}")
-            
-            name = cur.fetchall()
-            
-            cur.execute(f"DELETE FROM user WHERE user_id = {user_id}")
-            conn.commit()
-            
-            print(f"\nUser {name[0][0]} has been removed.\n")
-            
-            user_id = -1
-            break
-            
-        else:
-            print("\nUser not found.\n")
-
+            print(f"ID: {row[0]}, username: {row[1]}")
         
         conn.close()
+        
+        user_id = 0
+        
+        while user_id != -1: # -1 is used to cancel the operation.
+            user_found = False
+            
+            conn = sqlite3.connect("user_db")
+            cur = conn.cursor()
+            
+            result = conn.execute("SELECT * FROM user")
+            
+            user_id = int(input("\nInsert the ID of the user you'd like to remove. Insert -1 to cancel."))
+            
+            if user_id == -1:
+                break
+            
+            # finding the user
+            for row in result:
+                x = 0
+                
+                if user_id == row[x]:
+                    user_found = True
+                    break
+                x += 1
+                
+            if user_found:
+                result = cur.execute(f"SELECT username FROM user WHERE user_id = {user_id}")
+                
+                name = cur.fetchall()
+                
+                cur.execute(f"DELETE FROM user WHERE user_id = {user_id}")
+                conn.commit()
+                
+                print(f"\nUser {name[0][0]} has been removed.\n")
+                
+                user_id = -1
+                break
+                
+            else:
+                print("\nUser not found.\n")
+    
+            conn.close()
+    else:
+        print("\nOnly Admins can add or delete users.\n")
 
 def view_profile():
     conn = sqlite3.connect("user_db")
     cur = conn.cursor()
     
-    cur.execute(f"SELECT user_id, username, email FROM user WHERE user_id = {login[2]}")
+    cur.execute(f"SELECT user_id, username, email, is_admin FROM user WHERE user_id = {login[2]}")
     result = cur.fetchall()
     
     print("\nProfile:")
     print(f"ID: {result[0][0]}")
     print(f"Username: {result[0][1]}")
-    print(f"E-mail: {result[0][2]}\n")
+    print(f"E-mail: {result[0][2]}")
+    print(f"Is Admin: {result[0][3]}\n")
     
     conn.close()
 
@@ -163,6 +186,7 @@ while choice != 5:
                 
             case 3:
                 print("\nQuitting..")
+                break
         
     else:
         choice = int(input(f"Hello, {login[0]}.\nSelect an option.\n1. Add user\n2. View all users\n3. Delete user\n4. View profile\n5. Quit\n"))
